pyproject.toml
[project]
name = "nanobot-runner"
version = "0.17.0"
description = "Nanobot Runner - 桌面端私人AI跑步助理"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nanobot-ai>=0.1.5.post2",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "polars>=0.20.0",
    "pyarrow>=14.0.0",
    "fitparse>=1.1.0",
    "psutil>=7.2.2",
    "numpy>=2.4.2",
    "questionary>=1.10.0",
    "pydantic-settings>=2.0.0",
    "dulwich>=0.21.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "pytest-timeout>=2.2.0",
    "pytest-asyncio>=0.23.0",
]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.0.0,<2.0.0",
    "bandit>=1.7.0,<2.0.0",
    "safety>=2.0.0,<3.0.0",
    "pre-commit>=3.0.0,<4.0.0",
    "types-requests>=2.33.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
nanobotrun = "src.cli.app:app"

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = false

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
pythonpath = ["."]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html --cov-report=xml --timeout=30 --timeout-method=thread"
# 注册自定义标记
markers = [
    "e2e: 端到端测试",
    "slow: 慢速测试（执行时间>5 秒）",
    "performance: 性能测试",
    "integration: 集成测试",
    "unit: 单元测试",
]
# 超时配置
timeout = 30
timeout_method = "thread"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip without explicit strict parameter
    "F811",  # redefinition of unused variable
    "ARG001",  # unused function argument
    "ARG002",  # unused method argument
    "B904",  # raise from err
    "B007",  # loop control variable not used
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["SIM117", "ARG001", "ARG002", "ARG005", "F841", "E402", "E721", "B017"]  # 测试文件忽略嵌套with、未使用参数、类型比较、盲目异常等

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
disallow_incomplete_defs = false
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true

[tool.bandit]
skips = ["B101", "B601"]
targets = ["src"]
recursive = true
exclude_dirs = ["tests", "docs"]

[tool.safety]
output = "json"
ignore = []
